# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MoneyTrack personal savings and resale tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Goal {
  id: string;
  name: string;
  targetAmount: number;
  savedAmount: number;
  /** @nullable */
  targetDate: string | null;
  color: string;
  createdAt: string;
}

export interface GoalInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @exclusiveMinimum 0 */
  targetAmount: number;
  /** @nullable */
  targetDate?: string | null;
  /** @maxLength 30 */
  color?: string;
}

export interface GoalUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name?: string;
  /** @exclusiveMinimum 0 */
  targetAmount?: number;
  /** @nullable */
  targetDate?: string | null;
  /** @maxLength 30 */
  color?: string;
}

export interface Contribution {
  id: string;
  goalId: string;
  amount: number;
  date: string;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface ContributionInput {
  /** @exclusiveMinimum 0 */
  amount: number;
  date: string;
  /**
     * @maxLength 250
     * @nullable
     */
  note?: string | null;
}

export type ResaleItemStatus = typeof ResaleItemStatus[keyof typeof ResaleItemStatus];


export const ResaleItemStatus = {
  inventory: 'inventory',
  listed: 'listed',
  sold: 'sold',
  returned: 'returned',
} as const;

export interface ResaleItem {
  id: string;
  name: string;
  /** @nullable */
  category: string | null;
  /** @nullable */
  brand: string | null;
  /** @nullable */
  size: string | null;
  status: ResaleItemStatus;
  purchasePrice: number;
  purchaseDate: string;
  /** @nullable */
  listingPrice: number | null;
  /** @nullable */
  soldPrice: number | null;
  /** @nullable */
  platform: string | null;
  /** @nullable */
  platformFee: number | null;
  /** @nullable */
  paymentFee: number | null;
  /** @nullable */
  shippingCost: number | null;
  /** @nullable */
  otherCosts: number | null;
  /** @nullable */
  soldDate: string | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  netProfit: number | null;
  /** @nullable */
  profitMargin: number | null;
  createdAt: string;
}

export type ItemInputStatus = typeof ItemInputStatus[keyof typeof ItemInputStatus];


export const ItemInputStatus = {
  inventory: 'inventory',
  listed: 'listed',
  sold: 'sold',
  returned: 'returned',
} as const;

export interface ItemInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /**
     * @maxLength 80
     * @nullable
     */
  category?: string | null;
  /**
     * @maxLength 100
     * @nullable
     */
  brand?: string | null;
  /**
     * @maxLength 40
     * @nullable
     */
  size?: string | null;
  status?: ItemInputStatus;
  /** @minimum 0 */
  purchasePrice: number;
  purchaseDate: string;
  /**
     * @minimum 0
     * @nullable
     */
  listingPrice?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  soldPrice?: number | null;
  /**
     * @maxLength 80
     * @nullable
     */
  platform?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  platformFee?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  paymentFee?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  shippingCost?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  otherCosts?: number | null;
  /** @nullable */
  soldDate?: string | null;
  /**
     * @maxLength 2000
     * @nullable
     */
  notes?: string | null;
}

export type ItemUpdateStatus = typeof ItemUpdateStatus[keyof typeof ItemUpdateStatus];


export const ItemUpdateStatus = {
  inventory: 'inventory',
  listed: 'listed',
  sold: 'sold',
  returned: 'returned',
} as const;

export interface ItemUpdate {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name?: string;
  /**
     * @maxLength 80
     * @nullable
     */
  category?: string | null;
  /**
     * @maxLength 100
     * @nullable
     */
  brand?: string | null;
  /**
     * @maxLength 40
     * @nullable
     */
  size?: string | null;
  status?: ItemUpdateStatus;
  /** @minimum 0 */
  purchasePrice?: number;
  purchaseDate?: string;
  /**
     * @minimum 0
     * @nullable
     */
  listingPrice?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  soldPrice?: number | null;
  /**
     * @maxLength 80
     * @nullable
     */
  platform?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  platformFee?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  paymentFee?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  shippingCost?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  otherCosts?: number | null;
  /** @nullable */
  soldDate?: string | null;
  /**
     * @maxLength 2000
     * @nullable
     */
  notes?: string | null;
}

export interface MonthlyTotal {
  month: string;
  label: string;
  sales: number;
  profit: number;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  goal: 'goal',
  contribution: 'contribution',
  item: 'item',
  sale: 'sale',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  title: string;
  subtitle: string;
  amount: number;
  date: string;
}

export interface Dashboard {
  totalSaved: number;
  savingsTarget: number;
  goalCount: number;
  resaleProfit: number;
  monthlyProfit: number;
  activeListings: number;
  itemsSold: number;
  monthly: MonthlyTotal[];
  recentActivity: Activity[];
}

export interface Backup {
  formatVersion: number;
  exportedAt: string;
  goals: Goal[];
  contributions: Contribution[];
  items: ResaleItem[];
}

export interface BackupGoalInput {
  /** @minLength 1 */
  sourceId: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @exclusiveMinimum 0 */
  targetAmount: number;
  /** @nullable */
  targetDate?: string | null;
  /** @maxLength 30 */
  color?: string;
}

export interface BackupContributionInput {
  /** @minLength 1 */
  sourceGoalId: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  date: string;
  /**
     * @maxLength 250
     * @nullable
     */
  note?: string | null;
}

export interface BackupRestoreInput {
  goals: BackupGoalInput[];
  contributions: BackupContributionInput[];
  items: ItemInput[];
}

export interface BackupRestoreResult {
  goalsImported: number;
  contributionsImported: number;
  itemsImported: number;
}

export type GetItemsParams = {
status?: GetItemsStatus;
platform?: string;
search?: string;
};

export type GetItemsStatus = typeof GetItemsStatus[keyof typeof GetItemsStatus];


export const GetItemsStatus = {
  all: 'all',
  inventory: 'inventory',
  listed: 'listed',
  sold: 'sold',
  returned: 'returned',
} as const;