# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MoneyTrack personal savings and resale tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /backup/restore:
    post:
      operationId: restoreBackup
      tags: [dashboard]
      summary: Replace the signed-in user's data from a backup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupRestoreInput"
      responses:
        "200":
          description: Restored record counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupRestoreResult"
        "400":
          description: Invalid backup data
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List savings goals
      responses:
        "200":
          description: Savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{goalId}:
    parameters:
      - name: goalId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Goal not found
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a savings goal and its contribution history
      responses:
        "204":
          description: Goal deleted
        "404":
          description: Goal not found
  /goals/{goalId}/contributions:
    parameters:
      - name: goalId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getGoalContributions
      tags: [goals]
      summary: List contributions to a savings goal
      responses:
        "200":
          description: Goal contributions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contribution"
    post:
      operationId: createGoalContribution
      tags: [goals]
      summary: Add money to a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContributionInput"
      responses:
        "201":
          description: Added contribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contribution"
        "404":
          description: Goal not found
  /contributions/{contributionId}:
    parameters:
      - name: contributionId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteContribution
      tags: [goals]
      summary: Delete a contribution
      responses:
        "204":
          description: Contribution deleted
        "404":
          description: Contribution not found
...
components:
  schemas:
...
    Goal:
      type: object
      required: [id, name, targetAmount, savedAmount, targetDate, color, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        targetAmount: { type: number }
        savedAmount: { type: number }
        targetDate: { type: ["string", "null"] }
        color: { type: string }
        createdAt: { type: string, format: date-time }
    GoalInput:
      type: object
      required: [name, targetAmount]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        targetAmount: { type: number, exclusiveMinimum: 0 }
        targetDate: { type: ["string", "null"] }
        color: { type: string, maxLength: 30 }
    GoalUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        targetAmount: { type: number, exclusiveMinimum: 0 }
        targetDate: { type: ["string", "null"] }
        color: { type: string, maxLength: 30 }
    Contribution:
      type: object
      required: [id, goalId, amount, date, note, createdAt]
      properties:
        id: { type: string }
        goalId: { type: string }
        amount: { type: number }
        date: { type: string }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ContributionInput:
      type: object
      required: [amount, date]
      properties:
        amount: { type: number, exclusiveMinimum: 0 }
        date: { type: string }
        note: { type: ["string", "null"], maxLength: 250 }
...
    ItemInput:
      type: object
      required: [name, purchasePrice, purchaseDate]
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        category: { type: ["string", "null"], maxLength: 80 }
        brand: { type: ["string", "null"], maxLength: 100 }
        size: { type: ["string", "null"], maxLength: 40 }
        status:
          type: string
          enum: [inventory, listed, sold, returned]
        purchasePrice: { type: number, minimum: 0 }
        purchaseDate: { type: string }
        listingPrice: { type: ["number", "null"], minimum: 0 }
        soldPrice: { type: ["number", "null"], minimum: 0 }
        platform: { type: ["string", "null"], maxLength: 80 }
        platformFee: { type: ["number", "null"], minimum: 0 }
        paymentFee: { type: ["number", "null"], minimum: 0 }
        shippingCost: { type: ["number", "null"], minimum: 0 }
        otherCosts: { type: ["number", "null"], minimum: 0 }
        soldDate: { type: ["string", "null"] }
        notes: { type: ["string", "null"], maxLength: 2000 }
...
    BackupGoalInput:
      type: object
      required: [sourceId, name, targetAmount]
      properties:
        sourceId: { type: string, minLength: 1 }
        name: { type: string, minLength: 1, maxLength: 120 }
        targetAmount: { type: number, exclusiveMinimum: 0 }
        targetDate: { type: ["string", "null"] }
        color: { type: string, maxLength: 30 }
    BackupContributionInput:
      type: object
      required: [sourceGoalId, amount, date]
      properties:
        sourceGoalId: { type: string, minLength: 1 }
        amount: { type: number, exclusiveMinimum: 0 }
        date: { type: string }
        note: { type: ["string", "null"], maxLength: 250 }
    BackupRestoreInput:
      type: object
      required: [goals, contributions, items]
      properties:
        goals:
          type: array
          items:
            $ref: "#/components/schemas/BackupGoalInput"
        contributions:
          type: array
          items:
            $ref: "#/components/schemas/BackupContributionInput"
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemInput"
    BackupRestoreResult:
      type: object
      required: [goalsImported, contributionsImported, itemsImported]
      properties:
        goalsImported: { type: integer }
        contributionsImported: { type: integer }
        itemsImported: { type: integer }