# lib/api-zod/src/generated/types/backupContributionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MoneyTrack personal savings and resale tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface BackupContributionInput {
  /** @minLength 1 */
  sourceGoalId: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  date: string;
  /**
     * @maxLength 250
     * @nullable
     */
  note?: string | null;
}