# lib/api-zod/src/generated/types/backupRestoreResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MoneyTrack personal savings and resale tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface BackupRestoreResult {
  goalsImported: number;
  contributionsImported: number;
  itemsImported: number;
}