# lib/api-zod/src/generated/types/contribution.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MoneyTrack personal savings and resale tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface Contribution {
  id: string;
  goalId: string;
  amount: number;
  date: string;
  /** @nullable */
  note: string | null;
  createdAt: Date;
}