# lib/api-zod/src/generated/types/goalUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * MoneyTrack personal savings and resale tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name?: string;
  /** @exclusiveMinimum 0 */
  targetAmount?: number;
  /** @nullable */
  targetDate?: string | null;
  /** @maxLength 30 */
  color?: string;
}